Update Term_Project.R
I changed the name of the above variables because they were bothering me. Can you go through your code and change the names to appropriately fit the name changes? It should be featured throughout the factoring section. Also, in the factoring section, the reason the ggplot was not working was because you created factored variables that were already present as dummy variable. I already had all the mode choices as a 0,1 option, then you factored them to make 1,2.